Changed Points to Double
Now when user receives half points, they will not loose 0.5 points

+++ GoFetch/app/src/main/java/edu/western/cs/gofetch/lesson_related_activities/NavStartLesson.java
@@ -42,8 +42,8 @@ public class NavStartLesson extends AppCompatActivity implements NavigationView.
     private boolean timerRunning;
     private boolean timerFinished;
 
-    private int points;
-    private int addedPoints;
+    private double points;
+    private double addedPoints;
 
     Realm realm;
     @Override
